+++ src/main/java/com/github/jlangch/venice/util/excel/ExcelSheetReader.java
@@ -45,22 +45,22 @@ public class ExcelSheetReader {
     }
 
     public int getIndex() {
-    	return sheet.getIndex();
+    	return sheet.getIndex() + 1;  // 1-based
     }
 
     public int getFirstRowNum() {
         final int n = sheet.getFirstRowNum();
-        return n < 0 ? n : n + 1;
+        return n < 0 ? n : n + 1;  // 1-based
     }
 
     public int getLastRowNum() {
         final int n = sheet.getLastRowNum();
-        return n < 0 ? n : n + 1;
+        return n < 0 ? n : n + 1;  // 1-based
     }
 
     public int getFirstCellNum(final int row1) {
         final int n = sheet.getFirstCellNum(row1-1);
-        return n < 0 ? n : n + 1;
+        return n < 0 ? n : n + 1;  // 1-based
     }
 
     public int getLastCellNum(final int row1) {